almost a working test
